refactor: Organize pages into feature-based structure
- Move PromptSettings.tsx to features/settings/pages/ for better settings organization
- Move LandingPage.tsx and landing components to features/landing/ for cohesive feature grouping
- Move DisplayContent.tsx to features/content/pages/ for content-related functionality
- Update all import statements to reflect new file locations
- Update AppRoutes.tsx to use new import paths
- Maintain PrivacyPolicy.tsx in pages/ as standalone legal page per CLAUDE.md guidelines
- All tests pass with new structure
- Follows CLAUDE.md feature-based organization principles
