+++ src/app/AppRoutes.tsx
@@ -7,10 +7,10 @@ import Campaign from '../features/campaign/pages/Campaign';
 import Dashboard from '../features/dashboard/pages/Dashboard';
 import CampaignContent from '../features/campaign/pages/CampaignContent';
 import NewCampaigns from '../features/campaign/pages/NewCampaigns';
-import DisplayContent from '../pages/DisplayContent';
-import PromptsSettings from '../pages/PromptSettings';
+import DisplayContent from '../features/content/pages/DisplayContent';
+import PromptsSettings from '../features/settings/pages/PromptSettings';
 import { useAuth } from '../features/auth/context/AuthContext';
-import LandingPage from '../pages/LandingPage';
+import LandingPage from '../features/landing/pages/LandingPage';
 import AuthActions from '../features/auth/pages/AuthActions';
 import settings from '../shared/components/settings';
 import ManageUsers from '../features/users/pages/ManageUsers';
