+++ src/features/landing/components/HeroSection.tsx
