+++ src/features/landing/components/HowBanner.tsx
